+++ src/Hy3Layout.cpp
@@ -5,33 +5,37 @@
 
 #include <dlfcn.h>
 #include <hyprland/src/Compositor.hpp>
+#include <hyprland/src/desktop/state/FocusState.hpp>
 #include <hyprland/src/config/ConfigManager.hpp>
 #include <hyprland/src/desktop/DesktopTypes.hpp>
 #include <hyprland/src/desktop/Workspace.hpp>
+#include <hyprland/src/desktop/rule/Engine.hpp>
 #include <hyprland/src/managers/LayoutManager.hpp>
 #include <hyprland/src/managers/PointerManager.hpp>
 #include <hyprland/src/managers/SeatManager.hpp>
 #include <hyprland/src/managers/input/InputManager.hpp>
 #include <hyprland/src/plugins/PluginAPI.hpp>
 #include <hyprland/src/plugins/PluginSystem.hpp>
+#include <hyprland/src/xwayland/XWayland.hpp>
 #include <hyprutils/math/Vector2D.hpp>
 #include <ranges>
 
 #include "Hy3Layout.hpp"
 #include "Hy3Node.hpp"
-#include "SelectionHook.hpp"
 #include "TabGroup.hpp"
 #include "globals.hpp"
 #include "src/SharedDefs.hpp"
 #include "src/desktop/WLSurface.hpp"
 #include "src/desktop/Window.hpp"
+#include "src/desktop/rule/Rule.hpp"
+#include "src/desktop/types/OverridableVar.hpp"
 #include "src/devices/IPointer.hpp"
 
 PHLWORKSPACE workspace_for_action(bool allow_fullscreen) {
 	if (g_pLayoutManager->getCurrentLayout() != g_Hy3Layout.get()) return nullptr;
 
-	auto workspace = g_pCompositor->m_lastMonitor->m_activeSpecialWorkspace;
-	if (!valid(workspace)) workspace = g_pCompositor->m_lastMonitor->m_activeWorkspace;
+	auto workspace = Desktop::focusState()->monitor()->m_activeSpecialWorkspace;
+	if (!valid(workspace)) workspace = Desktop::focusState()->monitor()->m_activeWorkspace;
 
 	if (!valid(workspace)) return nullptr;
 	if (!allow_fullscreen && workspace->m_hasFullscreenWindow) return nullptr;
@@ -95,6 +99,7 @@ void Hy3Layout::onWindowCreatedTiling(PHLWINDOW window, eDirection) {
 	});
 
 	this->insertNode(this->nodes.back());
+	window->m_workspace->updateWindows();
 }
 
 void Hy3Layout::insertNode(Hy3Node& node) {
@@ -140,7 +145,7 @@ void Hy3Layout::insertNode(Hy3Node& node) {
 	}
 
 	if (opening_after == nullptr) {
-		auto last_window = g_pCompositor->m_lastWindow;
+		auto last_window = Desktop::focusState()->window();
 		if (last_window != nullptr && last_window->m_workspace == node.workspace
 		    && !last_window->m_isFloating
 		    && (node.data.is_window() || last_window != node.data.as_window())
@@ -296,7 +301,7 @@ void Hy3Layout::onWindowRemovedTiling(PHLWINDOW window) {
 	    (uintptr_t) node->parent
 	);
 
-	window->unsetWindowData(PRIORITY_LAYOUT);
+	window->m_ruleApplicator->resetProps(Desktop::Rule::RULE_PROP_ALL, Desktop::Types::PRIORITY_LAYOUT);
 
 	if (window->isFullscreen()) {
 		g_pCompositor->setWindowFullscreenInternal(window, FSMODE_NONE);
@@ -331,6 +336,8 @@ void Hy3Layout::onWindowRemovedTiling(PHLWINDOW window) {
 				target_parent->recalcSizePosRecursive();
 		}
 	}
+
+	window->m_workspace->updateWindows();
 }
 
 void Hy3Layout::onWindowFocusChange(PHLWINDOW window) {
@@ -380,6 +387,13 @@ void Hy3Layout::recalculateMonitor(const MONITORID& monitor_id) {
 
 		top_node->recalcSizePosRecursive();
 	}
+
+#ifndef NO_XWAYLAND
+	CBox box = g_pCompositor->calculateX11WorkArea();
+	if (!g_pXWayland || !g_pXWayland->m_wm)
+			return;
+	g_pXWayland->m_wm->updateWorkArea(box.x, box.y, box.w, box.h);
+#endif
 }
 
 void Hy3Layout::recalculateWindow(PHLWINDOW window) {
@@ -399,7 +413,7 @@ ShiftDirection reverse(ShiftDirection direction) {
 }
 
 void Hy3Layout::resizeActiveWindow(const Vector2D& delta, eRectCorner corner, PHLWINDOW pWindow) {
-	auto window = pWindow ? pWindow : g_pCompositor->m_lastWindow.lock();
+	auto window = pWindow ? pWindow : Desktop::focusState()->window();
 	if (!valid(window)) return;
 
 	auto* node = this->getNodeFromWindow(window.get());
@@ -494,9 +508,6 @@ void Hy3Layout::fullscreenRequestForWindow(
 
 	const auto& monitor = window->m_monitor;
 
-	window->updateDynamicRules();
-	window->updateWindowDecos();
-
 	if (target_mode == FSMODE_NONE) {
 		auto* node = this->getNodeFromWindow(window.get());
 
@@ -508,7 +519,8 @@ void Hy3Layout::fullscreenRequestForWindow(
 			*window->m_realPosition = window->m_lastFloatingPosition;
 			*window->m_realSize = window->m_lastFloatingSize;
 
-			window->unsetWindowData(PRIORITY_LAYOUT);
+			window->m_ruleApplicator->resetProps(Desktop::Rule::RULE_PROP_ALL, Desktop::Types::PRIORITY_LAYOUT);
+			window->updateWindowData();
 		}
 	} else {
 		// save position and size if floating
@@ -581,7 +593,17 @@ std::any Hy3Layout::layoutMessage(SLayoutMessageHeader header, std::string conte
 	return "";
 }
 
-SWindowRenderLayoutHints Hy3Layout::requestRenderHints(PHLWINDOW window) { return {}; }
+SWindowRenderLayoutHints Hy3Layout::requestRenderHints(PHLWINDOW window) {
+	if (this->shouldRenderSelected(window.get())) {
+    static auto active_color = CConfigValue<Hyprlang::CUSTOMTYPE>("general:col.active_border");
+		return {
+			.isBorderGradient = true,
+			.borderGradient = static_cast<CGradientValueData*>((active_color.ptr())->getData()),
+		};
+	}
+
+	return {};
+}
 
 void Hy3Layout::switchWindows(PHLWINDOW pWindowA, PHLWINDOW pWindowB) {
 	// todo
@@ -638,7 +660,7 @@ PHLWINDOW Hy3Layout::findFloatingWindowCandidate(const CWindow* from) {
 	for (auto& w: g_pCompositor->m_windows | std::views::reverse) {
 		if (w->m_isMapped && !w->isHidden() && w->m_isFloating && !w->isX11OverrideRedirect()
 		    && w->m_workspace == from->m_workspace && !w->m_X11ShouldntFocus
-		    && !w->m_windowData.noFocus.valueOrDefault() && w.get() != from)
+		    && !w->m_ruleApplicator->noFocus().valueOrDefault() && w.get() != from)
 		{
 			return w;
 		}
@@ -688,8 +710,6 @@ void Hy3Layout::onEnable() {
 
 	mouseButtonPtr =
 	    HyprlandAPI::registerCallbackDynamic(PHANDLE, "mouseButton", &Hy3Layout::mouseButtonHook);
-
-	selection_hook::enable();
 }
 
 void Hy3Layout::onDisable() {
@@ -698,7 +718,6 @@ void Hy3Layout::onDisable() {
 	urgentHookPtr.reset();
 	tickHookPtr.reset();
 	mouseButtonPtr.reset();
-	selection_hook::disable();
 
 	for (auto& node: this->nodes) {
 		if (node.data.is_window()) {
@@ -925,7 +944,7 @@ void Hy3Layout::shiftFocus(
     bool visible,
     bool warp
 ) {
-	auto current_window = g_pCompositor->m_lastWindow.lock();
+	auto current_window = Desktop::focusState()->window();
 
 	if (current_window != nullptr) {
 		if (current_window->m_workspace->m_hasFullscreenWindow) {
@@ -938,7 +957,7 @@ void Hy3Layout::shiftFocus(
 
 			if (next_window != nullptr) {
 				g_pInputManager->unconstrainMouse();
-				g_pCompositor->focusWindow(next_window);
+				Desktop::focusState()->fullWindowFocus(next_window);
 				if (warp) Hy3Layout::warpCursorToBox(next_window->m_position, next_window->m_size);
 			}
 			return;
@@ -971,7 +990,7 @@ Hy3Node* Hy3Layout::focusMonitor(ShiftDirection direction) {
 
 	if (next_monitor) {
 		bool found = false;
-		g_pCompositor->setActiveMonitor(next_monitor);
+		Desktop::focusState()->rawMonitorFocus(next_monitor);
 		auto next_workspace = next_monitor->m_activeWorkspace;
 
 		if (next_workspace) {
@@ -999,7 +1018,7 @@ bool Hy3Layout::shiftMonitor(Hy3Node& node, ShiftDirection direction, bool follo
 	auto next_monitor = g_pCompositor->getMonitorInDirection(getShiftDirectionChar(direction));
 
 	if (next_monitor) {
-		g_pCompositor->setActiveMonitor(next_monitor);
+		Desktop::focusState()->rawMonitorFocus(next_monitor);
 		auto next_workspace = next_monitor->m_activeWorkspace;
 		if (next_workspace) {
 			moveNodeToWorkspace(node.workspace.get(), next_workspace->m_name, follow, false);
@@ -1010,7 +1029,7 @@ bool Hy3Layout::shiftMonitor(Hy3Node& node, ShiftDirection direction, bool follo
 }
 
 void Hy3Layout::toggleFocusLayer(const CWorkspace* workspace, bool warp) {
-	auto current_window = g_pCompositor->m_lastWindow.lock();
+	auto current_window = Desktop::focusState()->window();
 	if (!current_window) return;
 
 	PHLWINDOW target;
@@ -1022,7 +1041,7 @@ void Hy3Layout::toggleFocusLayer(const CWorkspace* workspace, bool warp) {
 
 	if (!target) return;
 
-	g_pCompositor->focusWindow(target);
+	Desktop::focusState()->fullWindowFocus(target);
 
 	if (warp) {
 		Hy3Layout::warpCursorWithFocus(target->middle());
@@ -1030,7 +1049,7 @@ void Hy3Layout::toggleFocusLayer(const CWorkspace* workspace, bool warp) {
 }
 
 void Hy3Layout::warpCursor() {
-	auto current_window = g_pCompositor->m_lastWindow.lock();
+	auto current_window = Desktop::focusState()->window();
 
 	if (current_window != nullptr) {
 		if (current_window != nullptr) {
@@ -1038,7 +1057,7 @@ void Hy3Layout::warpCursor() {
 		}
 	} else {
 		auto* node =
-		    this->getWorkspaceFocusedNode(g_pCompositor->m_lastMonitor->m_activeWorkspace.get());
+		    this->getWorkspaceFocusedNode(Desktop::focusState()->monitor()->m_activeWorkspace.get());
 
 		if (node != nullptr) {
 			Hy3Layout::warpCursorWithFocus(node->position + node->size / 2);
@@ -1055,8 +1074,7 @@ void changeNodeWorkspaceRecursive(Hy3Node& node, PHLWORKSPACE workspace) {
 		window->moveToWorkspace(workspace);
 		window->m_monitor = workspace->m_monitor;
 		window->updateToplevel();
-		window->updateDynamicRules();
-		window->uncacheWindowDecos();
+		Desktop::Rule::ruleEngine()->updateAllRules();
 	} else {
 		for (auto* child: node.data.as_group().children) {
 			changeNodeWorkspaceRecursive(*child, workspace);
@@ -1082,7 +1100,7 @@ void Hy3Layout::moveNodeToWorkspace(
 	if (origin == workspace.get()) return;
 
 	auto* node = this->getWorkspaceFocusedNode(origin);
-	auto focused_window = g_pCompositor->m_lastWindow.lock();
+	auto focused_window = Desktop::focusState()->window();
 	auto* focused_window_node = this->getNodeFromWindow(focused_window.get());
 
 	auto origin_ws = node != nullptr           ? node->workspace
@@ -1370,7 +1388,7 @@ void Hy3Layout::setNodeSwallow(const CWorkspace* workspace, SetSwallowOption opt
 }
 
 void Hy3Layout::killFocusedNode(const CWorkspace* workspace) {
-	auto last_window = g_pCompositor->m_lastWindow.lock();
+	auto last_window = Desktop::focusState()->window();
 	if (last_window != nullptr && last_window->m_isFloating) {
 		g_pCompositor->closeWindow(last_window);
 	} else {
@@ -1526,7 +1544,7 @@ bool Hy3Layout::shouldRenderSelected(const CWindow* window) {
 	auto* focused = root->getFocusedNode();
 	if (focused == nullptr
 	    || (focused->data.is_window()
-	        && focused->data.as_window() != g_pCompositor->m_lastWindow.lock()))
+	        && focused->data.as_window() != Desktop::focusState()->window()))
 		return false;
 
 	switch (focused->data.type()) {
@@ -1697,13 +1715,16 @@ void Hy3Layout::applyNodeDataToWindow(Hy3Node* node, bool no_animation) {
 		return;
 	}
 
-	window->unsetWindowData(PRIORITY_LAYOUT);
+	window->m_ruleApplicator->resetProps(Desktop::Rule::RULE_PROP_ALL, Desktop::Types::PRIORITY_LAYOUT);
+	window->updateWindowData();
 
 	auto nodeBox = CBox(node->position, node->size);
 	nodeBox.round();
 
 	window->m_size = nodeBox.size();
 	window->m_position = nodeBox.pos();
+
+	window->updateWindowDecos();
 
 	auto only_node = root_node != nullptr && root_node->data.as_group().children.size() == 1
 	              && root_node->data.as_group().children.front()->data.is_window();
@@ -1712,31 +1733,17 @@ void Hy3Layout::applyNodeDataToWindow(Hy3Node* node, bool no_animation) {
 	    && ((*no_gaps_when_only != 0 && (only_node || window->isFullscreen()))
 	        || window->isEffectiveInternalFSMode(FSMODE_FULLSCREEN)))
 	{
-		window->m_windowData.decorate = CWindowOverridableVar(
-		    true,
-		    PRIORITY_LAYOUT
-		); // a little curious but copying what dwindle does
-		window->m_windowData.noBorder =
-		    CWindowOverridableVar(*no_gaps_when_only != 2, PRIORITY_LAYOUT);
-		window->m_windowData.noRounding = CWindowOverridableVar(true, PRIORITY_LAYOUT);
-		window->m_windowData.noShadow = CWindowOverridableVar(true, PRIORITY_LAYOUT);
-
-		window->updateWindowDecos();
 
 		const auto reserved = window->getFullWindowReservedArea();
 
 		*window->m_realPosition = window->m_position + reserved.topLeft;
 		*window->m_realSize = window->m_size - (reserved.topLeft + reserved.bottomRight);
-
-		window->sendWindowSize(true);
 	} else {
 		auto reserved = window->getFullWindowReservedArea();
 		auto wb = node->getStandardWindowArea({-reserved.topLeft, -reserved.bottomRight});
 
 		*window->m_realPosition = wb.pos();
 		*window->m_realSize = wb.size();
-
-		window->sendWindowSize(true);
 
 		if (no_animation) {
 			g_pHyprRenderer->damageWindow(window);
@@ -1749,6 +1756,8 @@ void Hy3Layout::applyNodeDataToWindow(Hy3Node* node, bool no_animation) {
 
 		window->updateWindowDecos();
 	}
+
+	window->m_workspace->updateWindows();
 }
 
 bool shiftIsForward(ShiftDirection direction) {